user model init
