+++ luminous/app.js
@@ -3,10 +3,8 @@ require("dotenv").config();
 const express = require("express");
 const path = require("path");
 const connection = require("./utils/init.js");
-const Listing = require("./models/listing.js");
 const listingRouter = require("./routes/listing.js");
-const User = require("./models/user.js");
-const { randomUUID } = require("crypto");
+const userRouter = require("./routes/user.js");
 const sessions = require("./utils/sessions.js");
 
 const app = express();
@@ -28,64 +26,7 @@ app.get("/", (req, res) => {
   res.status(200).redirect(`/listings`);
 });
 
-// sign up get requist
-app.get("/user/signup", (req, res) => {
-  res.status(200).render("signup.ejs", { title: "signup page!", user: null });
-});
-
-// sign up middleware
-app.use("/user/signup", async (req, res, next) => {
-  try {
-    const { username, email, password } = req.body;
-    const user1 = new User({
-      username,
-      email,
-      password,
-    });
-    await user1.save();
-    res.status(200).redirect("/user/signin");
-  } catch (err) {
-    next(err);
-  }
-});
-
-// sign up get requist
-app.get("/user/signin", (req, res) => {
-  res.status(200).render("signin.ejs", { title: "signin page!", user: null });
-});
-
-// sign in middleware
-app.use("/user/signin", async (req, res, next) => {
-  try {
-    const { username, password } = req.body;
-    const user = await User.findOne({ username: username });
-    if (!user) {
-      res.send("Username is Incorrect!!");
-    }
-    if (user.password != password) {
-      res.send("Password is Incorrect!!");
-    }
-    const sessionId = randomUUID();
-    sessions[sessionId] = { username, id: user._id };
-    res.cookie("_session_uuid", sessionId, {
-      secure: true,
-      httpOnly: true,
-    });
-    res.status(200).redirect("/listings");
-  } catch (err) {
-    next(err);
-  }
-});
-
-// sign out requist
-app.get("/user/signout", (req, res) => {
-  res.cookie("_session_uuid", null, {
-    secure: true,
-    httpOnly: true,
-  });
-  res.status(200).redirect("/user/signin");
-});
-
+// user is logged in or not check.
 const isLogInUser = (req, res, next) => {
   if (!sessions[req.headers?.cookie?.split("=")[1]]) {
     res.status(401).send("session expired. login again!!");
@@ -95,8 +36,11 @@ const isLogInUser = (req, res, next) => {
   return next();
 };
 
+// route middleware
+app.use("/user", userRouter);
 app.use("/listings", isLogInUser, listingRouter);
 
+// err middleware
 app.use((err, req, res, next) => {
   const { status = 500, message } = err;
   res.status(status).send(message);
