+++ luminous/views/listings.ejs
@@ -1,6 +1,6 @@
 <%- include("partials/boilerplate", { title, user }); %>
   <h3 class="heading">
-    Luminious Listings!!
+    Luminious Listings for <%= user.username %>!!
   </h3>
   <div class="container">
     <% for(listing of listings) { %>
